django-api/requirements.txt
Django==4.2.6
psycopg2_binary==2.9.9
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt
drf-nested-routers
pytest-django